Add class to widget's UL for easier CSS
To allow this widget to be styled differently than other widgets it
needs a class attribute to identify it with CSS. This commit adds the
"wp-most-popular" class to the UL element.

No class is added to the LI elements because they can be selected as
children of the UL element in the style sheet.